rename existing to legacy in ExistingOpenApiIntegration folder and namespace
