+++ test/OpenApiTests/LegacyOpenApiIntegration/LegacyIntegrationDbContext.cs
@@ -4,16 +4,16 @@ using Microsoft.EntityFrameworkCore;
 // @formatter:wrap_chained_method_calls chop_always
 // @formatter:keep_existing_linebreaks true
 
-namespace OpenApiTests.ExistingOpenApiIntegration
+namespace OpenApiTests.LegacyOpenApiIntegration
 {
     [UsedImplicitly(ImplicitUseTargetFlags.Members)]
-    public sealed class ExistingIntegrationDbContext : DbContext
+    public sealed class LegacyIntegrationDbContext : DbContext
     {
         public DbSet<Airplane> Airplanes { get; set; }
         public DbSet<Flight> Flights { get; set; }
         public DbSet<FlightAttendant> FlightAttendants { get; set; }
 
-        public ExistingIntegrationDbContext(DbContextOptions<ExistingIntegrationDbContext> options)
+        public LegacyIntegrationDbContext(DbContextOptions<LegacyIntegrationDbContext> options)
             : base(options)
         {
         }
