+++ test/OpenApiTests/LegacyOpenApiIntegration/swagger.json
