+++ test/OpenApiTests/OpenApiTests.csproj
@@ -29,7 +29,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <OpenApiReference Include="ExistingOpenApiIntegration\swagger.json">
+    <OpenApiReference Include="LegacyOpenApiIntegration\swagger.json">
       <Namespace>OpenApiTests.ClientLibrary.GeneratedCode</Namespace>
       <ClassName>OpenApiClient</ClassName>
       <CodeGenerator>NSwagCSharp</CodeGenerator>
@@ -38,7 +38,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="ExistingOpenApiIntegration\swagger.json" />
+    <EmbeddedResource Include="LegacyOpenApiIntegration\swagger.json" />
   </ItemGroup>
 
   <!-- Fixes IntelliSense errors on openapi.json in Visual Studio 2019, which uses the schema for OpenAPI 3.1 by default. -->
